split unknown_0B8920 into unknown_0C4800
