+++ asm/non_matchings/unknown_0B8920/func_800C45A4.s
@@ -1,3 +1,8 @@
+.late_rodata
+glabel jpt_800E9360
+.word L800C4A88, L800C4A78, L800C4AD0, L800C4AE8, L800C4ADC, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4AE8, L800C4A6C
+
+.text
 glabel func_800C45A4
 /* 0C51A4 800C45A4 27BDFF78 */  addiu $sp, $sp, -0x88
 /* 0C51A8 800C45A8 AFB50028 */  sw    $s5, 0x28($sp)
@@ -308,19 +313,22 @@ glabel func_800C45A4
 /* 0C5648 800C4A48 2DE10018 */  sltiu $at, $t7, 0x18
 /* 0C564C 800C4A4C 10200026 */  beqz  $at, .L800C4AE8
 /* 0C5650 800C4A50 000F7880 */   sll   $t7, $t7, 2
-/* 0C5654 800C4A54 3C01800F */  lui   $at, %hi(D_800E9360) # $at, 0x800f
+/* 0C5654 800C4A54 3C01800F */  lui   $at, %hi(jpt_800E9360) # $at, 0x800f
 /* 0C5658 800C4A58 002F0821 */  addu  $at, $at, $t7
-/* 0C565C 800C4A5C 8C2F9360 */  lw    $t7, %lo(D_800E9360)($at)
+/* 0C565C 800C4A5C 8C2F9360 */  lw    $t7, %lo(jpt_800E9360)($at)
 /* 0C5660 800C4A60 00000000 */  nop   
 /* 0C5664 800C4A64 01E00008 */  jr    $t7
 /* 0C5668 800C4A68 00000000 */   nop   
+glabel L800C4A6C
 /* 0C566C 800C4A6C 956E0024 */  lhu   $t6, 0x24($t3)
 /* 0C5670 800C4A70 10000050 */  b     .L800C4BB4
 /* 0C5674 800C4A74 020E8021 */   addu  $s0, $s0, $t6
+glabel L800C4A78
 /* 0C5678 800C4A78 95790022 */  lhu   $t9, 0x22($t3)
 /* 0C567C 800C4A7C 86500020 */  lh    $s0, 0x20($s2)
 /* 0C5680 800C4A80 1000004C */  b     .L800C4BB4
 /* 0C5684 800C4A84 03F9F821 */   addu  $ra, $ra, $t9
+glabel L800C4A88
 /* 0C5688 800C4A88 864F0020 */  lh    $t7, 0x20($s2)
 /* 0C568C 800C4A8C 95620026 */  lhu   $v0, 0x26($t3)
 /* 0C5690 800C4A90 020F7023 */  subu  $t6, $s0, $t7
@@ -341,13 +349,16 @@ glabel func_800C45A4
 /* 0C56C4 800C4AC4 03198023 */  subu  $s0, $t8, $t9
 /* 0C56C8 800C4AC8 1000003A */  b     .L800C4BB4
 /* 0C56CC 800C4ACC 00000000 */   nop   
+glabel L800C4AD0
 /* 0C56D0 800C4AD0 956F0022 */  lhu   $t7, 0x22($t3)
 /* 0C56D4 800C4AD4 10000037 */  b     .L800C4BB4
 /* 0C56D8 800C4AD8 03EFF821 */   addu  $ra, $ra, $t7
+glabel L800C4ADC
 /* 0C56DC 800C4ADC 86500020 */  lh    $s0, 0x20($s2)
 /* 0C56E0 800C4AE0 10000034 */  b     .L800C4BB4
 /* 0C56E4 800C4AE4 00000000 */   nop   
 .L800C4AE8:
+glabel L800C4AE8
 /* 0C56E8 800C4AE8 956E0024 */  lhu   $t6, 0x24($t3)
 /* 0C56EC 800C4AEC 10000031 */  b     .L800C4BB4
 /* 0C56F0 800C4AF0 020E8021 */   addu  $s0, $s0, $t6
